Iterate group-field-history
This now applies to all controls in the container it's in, typically a
section. Ticking the box will set all controls to their historic value
(presumed to exist) and set the unchanged meta attr on their binding
node (even if they are not history aware). Unticking clears the value
and removes the attr.